refactor: [#45] extract module for pytest namespaces patch
